prediction_app/scripts/requirements.txt
# 数据导入脚本依赖
pandas>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
